requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
segment-anything-py>=0.0.1
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
filterpy>=1.4.5
deep-sort-realtime>=1.3.2
bytetrack>=0.3.0
pillow>=9.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
scipy>=1.10.0